Second Day
Still in process of implementing the challenges.

Parallax issues:  I have no idea how to do it from scratch.

The top left button is NOT supposed to be part of nav bar as it will be
fixed, and will need to make a slide-in..not sure if CSS-only is option.

Many left to go…